1) добавил явно debug версию приложения. В ней отличается название приложения в меню 2) добавил версии free и paid . В версии free задизейблин тогл переключения языков фильмов во фрагменте настроек 3) перенес пароли в файл, который игнорится гитом. Сделал соответствующие изменения в gradle файле
