+++ .gitignore
@@ -8,3 +8,5 @@
 .externalNativeBuild
 .cxx
 local.properties
+app/keys_ms
+signing.properties
