+++ app/src/debug/res/values/strings.xml
@@ -1,0 +1,31 @@
+<resources>
+    <string name="app_name">MovieSearch-debug</string>
+    <string name="bottom_menu_movies">Movies</string>
+    <string name="bottom_menu_favorites">Favorites</string>
+    <string name="bottom_menu_settings">Settings</string>
+
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="favorites_movies_fragment">Debug version</string>
+    <string name="movie_details">MovieDetails</string>
+    <string name="movie_list">MovieList</string>
+    <string name="russian_language">Russian language</string>
+    <string name="image_movie_poster">image_movie_poster</string>
+    <string name="reload">Reload</string>
+    <string name="error">Error</string>
+
+    <string name="menu_title_history">history</string>
+    <string name="menu_title_contacts">contacts</string>
+    <string name="note_hint">Введите заметку</string>
+    <string name="save_note">Сохранить</string>
+    <string name="google_maps">maps</string>
+
+    <string name="dialog_button_close">Закрыть</string>
+    <string name="dialog_title_no_gps">Доступ к геолокации закрыт</string>
+    <string name="dialog_message_no_gps">Разрешите доступ к геолокации, если хотите
+найти себя на карте</string>
+    <string name="default_address">Red Square, 1, Moscow</string>
+    <string name="button_search_map">Search</string>
+    <string name="marker_start">Start</string>
+    <string name="set_address">Введите адрес</string>
+    <string name="search_address">Найти</string>
+</resources>
